Autocenter map and update README
