+++ client/components/MapContainer.jsx
@@ -1,38 +1,73 @@
 import React from 'react';
-import { withScriptjs, withGoogleMap, GoogleMap, Marker, Circle, BicyclingLayer } from "react-google-maps";
+import GoogleMapReact from 'google-map-react';
+let GoogleMapsAPIKEY =  require('../../config.js').GoogleMapsAPIKEY;
 
-// const transit = {
-//   "featureType": "transit.line",
-//   "elementType": "geometry",
-//   "stylers": [
-//     {
-//       "color": "#00007F"
-//     },
-//     {
-//       "lightness": 19
-//     }
-//   ]
-// }
 
-const MapContainer = withScriptjs(withGoogleMap((props) =>
-  <GoogleMap
-    defaultZoom={14}
-    defaultCenter={{lat: props.lat, lng: props.long}}
-    defaultOptions={{
-      minZoom: 3,
+const Marker = () => {
+  return <div className="marker"></div>
+}
+
+const createMapOptions = () => {
+  return (
+    {
+      minZoom: 12,
+      maxZoom: 15,
       fullscreenControl: false,
       zoomControl: false,
       streetViewControl: false,
       scaleControl: true,
       mapTypeControl: false
-    }}
-  >
-    {props.isMarkerShown && <Marker position={{lat: props.lat, lng: props.long}}/>}
-    <BicyclingLayer autoUpdate />
-  </GoogleMap>
-))
+    }
+  )
+}
+
+class MapContainer extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      center: {
+        lat: this.props.lat, 
+        lng: this.props.long
+      }
+    }
+  }
+
+  componentDidMount() {
+    this.setState({
+      center: {
+        lat: this.props.lat, 
+        lng: this.props.long
+      }
+    })
+  }
+
+  render() {
+    return (
+      <div style={{ height: `340px`, width: `595px` }} className="map">
+        <GoogleMapReact
+          bootstrapURLKeys={{key: GoogleMapsAPIKEY}}
+          center={{
+            lat: this.props.lat, 
+            lng: this.props.long
+          }}
+          zoom={14}
+          layerTypes={this.props.shouldShowTransit ? ['TransitLayer'] : []}
+          options={createMapOptions}
+        >
+          <Marker
+            lat={this.props.lat}
+            lng={this.props.long}
+          />
+        </GoogleMapReact>
+      </div>
+    )
+  }
+}
+
+MapContainer.defaultProps = { 
+  center: {lat: 51.515026, lng: -0.175809},
+  zoom: 14
+}
+
 
 export default MapContainer;
-
-
-// styles: [...administrative, ...landscape, ...poi, ...road, ...transit, ...water]
