+++ public/styles.css
@@ -1,4 +1,6 @@
 body#neighbApp {
+  margin: 10px;
+  max-width: 700px;
   font-size: 16px;
   font-family: 'Roboto', sans-serif;
   line-height: 22px;
